# contracts/v1_api_swagger.json
{
  "openapi": "3.0.1",
...
  "info": {
    "title": "CipherSwarm Agent API",
    "description": "The CipherSwarm Agent API is used to allow agents to connect to the CipherSwarm server.",
    "version": "1.3",
    "license": {
      "name": "Mozilla Public License Version 2.0",
      "url": "https://www.mozilla.org/en-US/MPL/2.0/"
    }
  },
  "servers": [
    {
      "url": "https://{defaultHost}",
      "description": "The production server",
      "variables": {
        "defaultHost": {
          "default": "www.example.com"
        }
      }
    },
    {
      "url": "http://{hostAddress}:{hostPort}",
      "description": "The insecure server",
      "variables": {
        "hostAddress": {
          "default": "localhost"
        },
        "hostPort": {
          "default": "8080"
        }
      }
    }
  ],
  "security": [
    {
      "bearer_auth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorObject": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The id of the agent"
          },
          "host_name": {
            "type": "string",
            "description": "The hostname of the agent"
          },
          "client_signature": {
            "type": "string",
            "description": "The signature of the client"
          },
          "state": {
            "type": "string",
            "description": "The state of the agent",
            "enum": [
              "pending",
              "active",
              "stopped",
              "error"
            ]
          },
          "operating_system": {
            "type": "string",
            "description": "The operating system of the agent"
          },
          "devices": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "The descriptive name of a GPU or CPU device."
            }
          },
          "advanced_configuration": {
            "$ref": "#/components/schemas/AdvancedAgentConfiguration"
          }
        },
        "required": [
          "id",
          "host_name",
          "client_signature",
          "operating_system",
          "devices",
          "state",
          "advanced_configuration"
        ]
      },
      "AdvancedAgentConfiguration": {
        "type": "object",
        "properties": {
          "agent_update_interval": {
            "type": "integer",
            "nullable": true,
            "description": "The interval in seconds to check for agent updates"
          },
          "use_native_hashcat": {
            "type": "boolean",
            "nullable": true,
            "description": "Use the hashcat binary already installed on the client system"
          },
          "backend_device": {
            "type": "string",
            "nullable": true,
            "description": "The device to use for hashcat, separated by commas"
          },
          "opencl_devices": {
            "type": "string",
            "nullable": true,
            "description": "The OpenCL device types to use for hashcat, separated by commas"
          },
          "enable_additional_hash_types": {
            "type": "boolean",
            "description": "Causes hashcat to perform benchmark-all, rather than just benchmark"
          }
        },
        "required": [
          "agent_update_interval",
          "use_native_hashcat",
          "backend_device",
          "enable_additional_hash_types"
        ]
      },
...
    }
  },
  "paths": {
    "/api/v1/client/agents/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "required": true,
          "description": "id"
        }
      ],
      "get": {
        "summary": "Gets an instance of an agent",
        "tags": [
          "Agents"
        ],
        "description": "Returns an agent",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "getAgent",
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "id": 2571,
                      "host_name": "weissnat",
                      "client_signature": "macOS, High Sierra (10.13)",
                      "operating_system": "linux",
                      "state": "active",
                      "devices": [
                        "cpu",
                        "GPU"
                      ],
                      "advanced_configuration": {
                        "agent_update_interval": 36,
                        "use_native_hashcat": false,
                        "backend_device": null,
                        "opencl_devices": null,
                        "enable_additional_hash_types": false
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Updates the agent",
        "tags": [
          "Agents"
        ],
        "description": "Updates an agent",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "updateAgent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "id": 2573,
                      "host_name": "stoltenberg",
                      "client_signature": "Linux, RHEL 6.10",
                      "operating_system": "linux",
                      "state": "active",
                      "devices": [
                        "cpu",
                        "GPU"
                      ],
                      "advanced_configuration": {
                        "agent_update_interval": 24,
                        "use_native_hashcat": false,
                        "backend_device": null,
                        "opencl_devices": null,
                        "enable_additional_hash_types": false
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "The id of the agent"
                  },
                  "host_name": {
                    "type": "string",
                    "description": "The hostname of the agent"
                  },
                  "client_signature": {
                    "type": "string",
                    "description": "The signature of the client"
                  },
                  "operating_system": {
                    "type": "string",
                    "description": "The operating system of the agent"
                  },
                  "devices": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "The descriptive name of a GPU or CPU device."
                    }
                  }
                },
                "required": [
                  "id",
                  "host_name",
                  "client_signature",
                  "operating_system",
                  "devices"
                ]
              },
              "examples": {
                "request_example_1": {
                  "summary": "A request example",
                  "value": {
                    "id": 2573,
                    "client_signature": "Linux, RHEL 6.10",
                    "enabled": true,
                    "last_ipaddress": "",
                    "last_seen_at": null,
                    "host_name": "stoltenberg",
                    "operating_system": "linux",
                    "token": "jwTK41L3PsXctHdcJL7Pt1zt",
                    "user_id": 12821,
                    "created_at": "2024-10-01T21:16:40.622-04:00",
                    "updated_at": "2024-10-01T21:16:40.622-04:00",
                    "devices": [
                      "cpu",
                      "GPU"
                    ],
                    "advanced_configuration": {
                      "agent_update_interval": 24,
                      "use_native_hashcat": false,
                      "backend_device": null,
                      "opencl_devices": null,
                      "enable_additional_hash_types": false
                    },
                    "state": "active",
                    "custom_label": null
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}